+++ R/ggline.R
@@ -135,6 +135,7 @@ ggline<- function(data, x, y, group = 1,
                   label = NULL, font.label = list(size = 11, color = "black"),
                   label.select = NULL, repel = FALSE, label.rectangle = FALSE,
                   show.line.label = FALSE,
+                  position = "identity",
                   ggtheme = theme_pubr(),
                   ...)
 {
@@ -155,7 +156,7 @@ ggline<- function(data, x, y, group = 1,
     add = add, add.params = add.params, error.plot = error.plot,
     label = label, font.label = font.label, label.select = label.select,
     repel = repel, label.rectangle = label.rectangle,
-    show.line.label = show.line.label, ggtheme = ggtheme, ...)
+    show.line.label = show.line.label, position = position, ggtheme = ggtheme, ...)
   if(!missing(data)) .opts$data <- data
   if(!missing(x)) .opts$x <- x
   if(!missing(y)) .opts$y <- y
@@ -196,6 +197,7 @@ ggline_core <- function(data, x, y, group = 1,
                   show.line.label = FALSE,
                   font.label = list(size = 11, color = "black"),
                   repel = FALSE, label.rectangle = FALSE,
+                  position = "identity",
                   ggtheme = theme_pubr(),
                       ...)
 {
@@ -206,7 +208,6 @@ ggline_core <- function(data, x, y, group = 1,
   error.plot = error.plot[1]
   plot_type <- match.arg(plot_type)
   if("none" %in% add) add <- "none"
-  position = "identity"
   grouping.vars <- intersect(c(x, color, linetype, group, facet.by), names(data))
   . <- NULL
 
@@ -246,7 +247,7 @@ ggline_core <- function(data, x, y, group = 1,
   #:::::::::::::::::::::::::::::::::::::::
   add.params <- add.params %>%
     .add_item(error.plot = error.plot,
-              position = "identity", p_geom = "geom_line")
+              position = position, p_geom = "geom_line")
 
   # First add geom if any
   p <- add.params %>%
@@ -281,7 +282,8 @@ ggline_core <- function(data, x, y, group = 1,
     p <- p +
     geom_exec(geom_point, data = data_sum,
                color = point.color, shape = shape,
-               size = 1.2+point.size, stroke = stroke)
+               size = 1.2+point.size, stroke = stroke,
+              position = position)
     # Adjust shape when ngroups > 6, to avoid ggplot warnings
     p <-.scale_point_shape(p, data_sum, shape)
   }